requirements.txt
fastapi==0.135.3
uvicorn==0.43.0
pandas==2.2.3
google-cloud-bigquery==3.41.0